+++ src/features/Modal/HaltModal/HaltModalComponent.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Button } from '../../Shared/Button'
+import { Button } from '../../Shared/Button/Button'
 
 const HaltModalComponent = ({
     hasGameEnded,
