+++ src/styles/main.scss
@@ -44,4 +44,4 @@ $fa-font-path: "../../node_modules/font-awesome/fonts";
 @import "../features/LandingPage/StaticPages/HowToPlay/HowToPlay.scss";
 @import "../features/LandingPage/StaticPages/News/News.scss";
 @import "../features/Timer/Timer.scss";
-@import '../features/Shared/Button.scss';
+@import '../features/Shared/Button/Button.scss';
